restructured resolvers to eliminate queryResolvers and mutationResolvers files, adding Query and Mutation objects to each resolver file containing respective resolvers
